CB-30359: Upgrade patch restriction service should use release version instead of CDH version
In this change, I've modified the UpgradePermissionProvider to pass the release version to the UpgradePathRestrictionService instead of the CDH runtime version if the release-version is present in the target image. Otherwise, the service will use the CDH runtime version as before.

The reason we introduced this change is that we should rely on the release version instead of the CDH version. This is because there may be cases where a new image is released with an updated CM but the same CDH version. In such situations, the release version will correctly indicate which versions should be used