CB-30370 Move the change image flow earlier in the Data Lake upgrade process
Data Lake upgrades will be more robust if we update the stack image
to the target image id, before the cluster upgrade. In the event of
a failed upgrade, a subsequent repair will bring up new nodes with
the new image.
For better code organization, change image has been added into the
core flow chain, but for backward compatibility, the change image
step was left in the datalake side flow. It can be remoced in the
next Cloudbreak version.