CB-30544 Adjust bootstrapping to handle different salt versions
Only the latest version of salt master can communicate to a minion with previous version, thus we need to drop checking all master to minion test.
Instead MinionAcceptor error handling is enriched. It will return with the master fqdn when there is a failure, also SaltBootstrap will populate
targets based on this error.