CB-30652 - Provider syncer incorrectly signals default outbound notification
I mistakenly expected `shouldUpgradeOutbound()` method to throw CloudbreakServiceException in case
`cloudResource.getParameter()` could not be parsed.
This will not be thrown as NetworkAttributes is annotated with `@JsonIgnoreProperties(ignoreUnknown = true)`
`cloudResource.getParameterStrict()` won't throw an exception either as
even though the strict mapper has `FAIL_ON_UNKNOWN_PROPERTIES` enabled, that annotation allows silently ignoring it.

This results in all fields in NetworkAttributes that are not present in the JSON become null,
returning an almost-null object, only attributes match amongst all the *Attributes classes.

Fix is to check if `networkAttributes.getNetworkId()` is not null to filter these out.