CB-30612 - New instances added to executor group have random mappings
When the volume management and disk handling was implemented, there was no
vertical scaling - disk add feature.
This means there were gaps in cases when there could be multiple VolumeSet
resources.
The symptom was the following:
1. vertical scale - add new disks to group
2. upscale group
3. new nodes volumes got mounted in an undeterministic order

Desired behaviour:
3. new nodes volumes should be mounted to preserve the ordering
 - original volumes first and additional volumes afterwards, keeping the ordering
of addition.

The commit contains 3 fixes were required to achieve this and have consistent mounting.