CB-30708 Map CM error for RangerAdmin command
Created a simple framework for mapping CM error commands based on an already existing solution, and extended it with RangerAdmin command failure in hybrid setup.
When setting up a hybrid connection, the hybrid environment's realm must be added to on-premises CM. Validating it is problematic, because PvC CP does not provide an API to retrieve the setting, and even if it is set in CM it requires service restart. So based on this, instead of validation, a new error mapper was added to point to the root cause of the error.