CB-30790 Adjust password generation to include special characters and be at least 32 chars
CM requires all password to be longer and has special characters in it.
Some quite safe special characters has been chosen and from those 10 is shuffled into the passwords.
This way the length will be 36 and all CM requirement is met.

Note: safe special characters are the less likely to cause issues in shell or jdbc URLs based on some research.