CB-31087 - [CDP][mow-prod] Cloudbreak Quartz threadpool active thread count high [freeipa][quartztaskexecutor]
stack-salt-status-checker-jobs are running for 424 seconds due to some excessive (15 times) retries.
The quick fix is to reduce the number of retries specifically for the causing commond execution, as
it can be parameterized due to a latest feature.

You can find more context here: https://cloudera.atlassian.net/browse/CB-31087?focusedCommentId=5804032