CB-31085: Fix GCP subnet ids to match the name instead of the internal id returned by GCP
When we are creating an environment, we correctly set the id of the subnets in the `environment_network` table to the name of the subnet. Then, during the environment creation, this id got overwritten by an internal id value returned by the GCP API. We would use this id from that point to refer to that subnet, which would lead to incorrect behavior, because our code would think it is a different subnet than what was originally identified during the environment creation.

This commit changes this behavior, so that we use the subnet name as the id, the same way we do in the case of Azure and AWS. This also fixes the link to the subnet under the network tab on the UI.

This commit also fixes the existing records in the environmentdb using a migration SQL. There are also affected records in the cbdb (and possibly freeipadb). Those can be fixed after the environmentdb has been fixed, as we can only retrieve the correct subnet id values from there. Possibly a stack patcher can be used to update the cbdb (and freeipadb if needed).