CB-30778 FlowTriggerCondition skip should skip only the actual flow and not the whole flowchain
Refactor the `Flow2Handler` class to pass the new `FlowEventContext` as a single argument instead of multiple variables, avoiding methods with 8+ parameters.

Fix the case where flow variables were not saved in the previous commit. Use `flow.getVariables()` between flow steps to save variables and use `contextParams` for the final flow step and between flows in a flowchain.