CB-27639 Poll environments2 API for usersync status when running mowdev tests
When fetching or waitin for a usersync status a fallback logic is implemented to call env2api when FMS returns with not found.
This integration is made seamless by adjusting FreeIPA/FMS client/Dtos to figure out if env2api call is required. This way our QA
tests doesn't require an WIAM to be running, and we don't have to adjust every test to the new setup.

environment2api client in our code got a fresh yaml from thunderhead, so the required models are generated. Gradle file is adjusted
to generate the client classes, also manipulates the endpoint in the yaml to ensure the generated `EnvironmentPublicApi` is not called
`DefaultApi`. This is done by adding tags to the endpoints in the yaml.