CB-9263 - Review FreeIPA notifications because it seems some of them are duplicated or not really useful
I will explicitly deny the structured event-based and flow-based FreeIPA notifications as they were full of duplicates, the messages did not make sense, and did not have any added value. This will be performed in a separate UI PR.

The only notifications presented on the UI will be of the “NOTIFICATION“ type, purposely emitted to provide more context around the current state of FreeIPA instances.

This commit contains the following
 - EventSender service implementation in FreeIPA module
 - re-enabling the structured event collection for FreeIPA
 - implementation of “FreeIPA create” flow notifications as a starter, and it can be extended with all the flows in FreeIPA later on.