CB-31205 Update GCP subnet IDs in the cbdb
The previous commit a588558ca5db67d148fea44c22dac013def8b094 fixed the issue, that we were using the GCP provider side ID for subnets, instead of the name, like we do for all other providers. Both new and existing the records in the environmentdb should have the correct values.

The incorrect values are also present in the cbdb, in three tables: network, instancegroupnetwork, resource. These also need to be corrected, otherwise when only comparing IDs of the subnets, CB won't know which provider side ID is referring to which known subnet name. This stack patcher collects the provider side IDs and pairs the with their name, then updates the cbdb using this mapping.