CB-31231: Add the ability to skip the highstate during a salt update
During certain scenarios, like when the embedded DB's certificate has expired, it is necessary to skip the highstate during the salt update phase of the secret rotation, since the highstate would try to reach CM, which is unavailable because of the expired cert. As workaround for these kind of problems, we can skip just the highstate part of the salt update flow. By default, the `skipHighstate` flag is always false, and it is only skipped if an explicit `true` value is passed in the flow's trigger event.