CB-31309 Fix timing issue in testClusterWithEmptyEnvironmentRequest
The FreeIPA was not yet created when awaitForFlow detected that it finished, because no flow was found for the resource.
Instead the test should wait for the AVAIABLE status, like how it is done in other places.
Modified the SDX creation await too to be the same as in AbstractIntegrationTest.
Also modified a duplicate log line in FreeIpaClientFactory to help debugging next time.