CB-31360 Fix patch upgrade calculation logic
During DataHub upgrades, when calculating if it is a patch upgrade or not, the code uses the CDH version from the DB, which in the case of the DataHubs, is not yet updated at this point, so the logic would always compare the current image to itself, instead of the target image, so every upgrade would be a patch upgrade. This has been fixed, by instead of using the version from the DB, we use the version coming from the request in the handler.

Also refactored the CDH version parsing util classes into a single `CdhVersionProvider` util class. Removed the `VersionNormalizer` util class.