CB-31136 Investigate retries in salt state running, we should avoid retrying fatal errors
The purpose of this effort is to lower the runtime of long-running status-checker Quartz jobs.

This commit contains the following to achieve this:
- a retry framework for running salt states with controlled, exception-based retries
- collected the most common non-retriable salt errors from prod Splunk and included these in the
initial implementation

Here are the number of occurrances in the last 30 days:

Failed to execute: {Comment=Command "/opt/salt/scripts/trust_status_validation.sh"	35249
Stderr=Configuration file could not be loaded.	620
Parameter validation failed:	40
Stdout=Failed to determine the current java version	23